fix: 修复类型转换错误并添加完整 Proto 定义
主要更改：
1. 修复 bytebuffer.go 中的类型转换错误（byte -> uint32）
2. 创建完整的 Protocol Buffers 消息定义（douyin.proto）
3. 临时禁用 WebView2 以避免 Windows SDK 依赖
4. 添加 WEBVIEW2_FIX.md 详细修复指南
5. 添加 BUILD_FIX_v2.md 修复报告

技术细节：
- ByteBuffer: 在位运算中为 c2, c3, c4 添加显式 uint32() 转换
- Proto: 包含所有抖音直播间消息类型（参考 dycast 和 DouyinBarrageGrab）
- WebView2: 可选功能，核心采集逻辑不受影响

测试状态：
- ✅ browser-monitor 打包成功
- ✅ server-active 编译成功
- 🟡 server-go 需要 Windows 环境或禁用 WebView2

参考：
- https://github.com/skmcj/dycast
- https://github.com/WanGuChou/DouyinBarrageGrab

+++ server-go/internal/parser/bytebuffer.go
@@ -182,7 +182,7 @@ func (bb *ByteBuffer) ReadString(length int) (string, error) {
 			}
 			c2 := bytes[i+1]
 			c3 := bytes[i+2]
-			if ((c2 | uint32(c3)<<8) & 0xc0c0) != 0x8080 {
+			if ((uint32(c2) | uint32(c3)<<8) & 0xc0c0) != 0x8080 {
 				result += "\uFFFD"
 				i++
 			} else {
@@ -203,7 +203,7 @@ func (bb *ByteBuffer) ReadString(length int) (string, error) {
 			c2 := bytes[i+1]
 			c3 := bytes[i+2]
 			c4 := bytes[i+3]
-			if ((c2 | uint32(c3)<<8 | uint32(c4)<<16) & 0xc0c0c0) != 0x808080 {
+			if ((uint32(c2) | uint32(c3)<<8 | uint32(c4)<<16) & 0xc0c0c0) != 0x808080 {
 				result += "\uFFFD"
 				i++
 			} else {
